fixup!

+++ src/multi/converters/converter_utils.cpp
@@ -121,7 +121,7 @@ namespace libp2p::multi::converters {
     const std::string hex = hex_upper(bytes);
 
     // Process Hex String
-    while (lastpos < bytes.size() * 2) {
+    while (lastpos < (size_t)bytes.size() * 2) {
       gsl::span<const uint8_t, -1> pid_bytes{bytes};
       int protocol_int = UVarint(pid_bytes.subspan(lastpos / 2)).toUInt64();
       Protocol const *protocol =
